requirements/base.txt
Django==5.0.3
python-environ==0.4.54
psycopg2-binary==2.9.9
django-ckeditor==6.7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
celery==5.4.0
django-mptt==0.16.0
